perf: use uint128 for Escrow amounts (saves 3 storage slots)
sharesHeld, principalAssetsSnapshot, committedAssets changed from
uint256 to uint128. Max uint128 ~3.4e38 is sufficient for any
practical share/asset amount. Reduces Escrow struct from ~11 to ~8
storage slots, saving ~66K gas per fundEscrow.

+++ contracts/stablecoin/v2/YieldEscrowV2.sol
@@ -69,13 +69,15 @@ contract YieldEscrowV2 is AccessControl, ReentrancyGuard {
         DisputeResolution disputeTimeoutResolution;
     }
 
+    /// @dev sharesHeld/principalAssetsSnapshot/committedAssets use uint128
+    ///      (max ~3.4e38 — sufficient for any practical share/asset amount)
     struct Escrow {
         address buyer;
         address merchant;
         address refundRecipient;
-        uint256 sharesHeld;
-        uint256 principalAssetsSnapshot;
-        uint256 committedAssets;
+        uint128 sharesHeld;
+        uint128 principalAssetsSnapshot;
+        uint128 committedAssets;
         uint40 releaseAfter;
         uint16 buyerBps;
         EscrowStatus status;
@@ -493,9 +495,9 @@ contract YieldEscrowV2 is AccessControl, ReentrancyGuard {
             buyer: buyer,
             merchant: merchant,
             refundRecipient: refundRecipient,
-            sharesHeld: sharesIn,
-            principalAssetsSnapshot: principalAssetsSnapshot,
-            committedAssets: committedAssets,
+            sharesHeld: uint128(sharesIn),
+            principalAssetsSnapshot: uint128(principalAssetsSnapshot),
+            committedAssets: uint128(committedAssets),
             releaseAfter: terms.releaseAfter,
             buyerBps: buyerBps,
             status: EscrowStatus.FUNDED,
